+++ src/main/java/dinkplugin/util/Utils.java
@@ -1,0 +1,46 @@
+package dinkplugin.util;
+
+import lombok.experimental.UtilityClass;
+import net.runelite.api.Client;
+import net.runelite.api.widgets.Widget;
+import net.runelite.api.widgets.WidgetInfo;
+import net.runelite.client.util.ColorUtil;
+import okhttp3.MediaType;
+import okhttp3.MultipartBody;
+import org.jetbrains.annotations.NotNull;
+
+import javax.imageio.ImageIO;
+import java.awt.Color;
+import java.awt.image.BufferedImage;
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+
+@UtilityClass
+public class Utils {
+
+    public final Color PINK = ColorUtil.fromHex("#f40098"); // analogous to RED in CIELCh_uv color space
+    public final Color RED = ColorUtil.fromHex("#ca2a2d"); // red used in pajaW
+
+    public boolean isSettingsOpen(@NotNull Client client) {
+        Widget widget = client.getWidget(WidgetInfo.SETTINGS_INIT);
+        return widget != null && !widget.isSelfHidden();
+    }
+
+    public String getPlayerName(Client client) {
+        return client.getLocalPlayer().getName();
+    }
+
+    public byte[] convertImageToByteArray(BufferedImage bufferedImage) throws IOException {
+        ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
+        ImageIO.write(bufferedImage, "png", byteArrayOutputStream);
+        return byteArrayOutputStream.toByteArray();
+    }
+
+    public boolean hasImage(@NotNull MultipartBody body) {
+        return body.parts().stream().anyMatch(part -> {
+            MediaType type = part.body().contentType();
+            return type != null && "image".equals(type.type());
+        });
+    }
+
+}